pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "microlog"
version = "0.1.0"
description = "Modern Python logging library with distributed tracing support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "MicroLog Team", email = "team@microlog.dev"}
]
maintainers = [
    {name = "MicroLog Team", email = "team@microlog.dev"}
]
keywords = [
    "logging",
    "async",
    "distributed-tracing",
    "trace-context",
    "microservices",
    "production-logging",
    "json-logging",
    "structured-logging"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Logging",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
dependencies = []

[project.optional-dependencies]
# Configuration file support
config = [
    "pyyaml>=6.0",
]

# Metrics and observability
metrics = [
    "prometheus-client>=0.17.0",
]

# All optional features
all = [
    "pyyaml>=6.0",
    "prometheus-client>=0.17.0",
]

# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/microlog"
Documentation = "https://github.com/yourusername/microlog/blob/main/README.md"
Repository = "https://github.com/yourusername/microlog"
Issues = "https://github.com/yourusername/microlog/issues"
Changelog = "https://github.com/yourusername/microlog/blob/main/CHANGELOG.md"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
microlog = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

